+++ pycosio/_core/functions_os_path.py
@@ -1,10 +1,11 @@
 # coding=utf-8
 """Cloud object compatibles standard library 'os.path' equivalent functions"""
 import os
-from os.path import relpath as os_path_relpath
+from os.path import relpath as os_path_relpath, samefile as os_path_samefile
 
 from pycosio._core.storage_manager import get_instance
-from pycosio._core.functions_core import equivalent_to
+from pycosio._core.functions_core import equivalent_to, format_and_is_storage
+from pycosio._core.exceptions import handle_os_exceptions
 
 
 @equivalent_to(os.path.getsize)
@@ -46,6 +47,23 @@ def getmtime(path):
     return get_instance(path).getmtime(path)
 
 
+@equivalent_to(os.path.isabs)
+def isabs(path):
+    """
+    Return True if path is an absolute pathname.
+
+    Equivalent to "os.path.isabs".
+
+    Args:
+        path (path-like object): File path or URL.
+
+    Returns:
+        bool: True if path is absolute.
+    """
+    # If detected as storage path, it is an absolute path.
+    return True
+
+
 @equivalent_to(os.path.isfile)
 def isfile(path):
     """
@@ -60,6 +78,22 @@ def isfile(path):
         bool: True if file exists.
     """
     return get_instance(path).isfile(path)
+
+
+@equivalent_to(os.path.ismount)
+def ismount(path):
+    """
+    Return True if pathname path is a mount point.
+
+    Equivalent to "os.path.ismount".
+
+    Args:
+        path (path-like object): File path or URL.
+
+    Returns:
+        bool: True if path is a mount point.
+    """
+    return True if not get_instance(path).relpath(path) else False
 
 
 @equivalent_to(os.path.relpath)
@@ -87,3 +121,68 @@ def relpath(path, start=None):
         # Replaces "\" by "/" for Windows.
         return os_path_relpath(relative, start=start).replace('\\', '/')
     return relative
+
+
+def samefile(path1, path2):
+    """
+    Return True if both pathname arguments refer to the same file or directory.
+
+    Equivalent to "os.path.samefile".
+
+    Args:
+        path1 (path-like object): File path or URL.
+        path2 (path-like object): File path or URL.
+
+    Returns:
+        bool: True if same file or directory.
+    """
+    # Handles path-like objects and checks if storage
+    path1, path1_is_storage = format_and_is_storage(path1)
+    path2, path2_is_storage = format_and_is_storage(path2)
+
+    # Local files: Redirects to "os.path.samefile"
+    if not path1_is_storage and not path2_is_storage:
+        return os_path_samefile(path1, path2)
+
+    # One path is local, the other storage
+    if not path1_is_storage or not path2_is_storage:
+        return False
+
+    with handle_os_exceptions():
+        # Paths don't use same storage
+        system = get_instance(path1)
+        if system is not get_instance(path2):
+            return False
+
+        # Relative path are different
+        elif system.relpath(path1) != system.relpath(path2):
+            return False
+
+    # Same files
+    return True
+
+
+@equivalent_to(os.path.splitdrive)
+def splitdrive(path):
+    """
+    Split the pathname path into a pair (drive, tail) where drive is either a
+    mount point or the empty string. On systems which do not use drive
+    specifications, drive will always be the empty string.
+
+    In all cases, drive + tail will be the same as path.
+
+    Equivalent to "os.path.splitdrive".
+
+    Args:
+        path (path-like object): File path or URL.
+
+    Returns:
+        tuple of str: drive, tail.
+    """
+    relative = get_instance(path).relpath(path)
+    drive = path.rsplit(relative, maxsplit=1)[0]
+    if drive and not drive[-2:] == '//':
+        # Keep "/" tail side
+        relative = '/' + relative
+        drive = drive.rstrip('/')
+    return drive, relative